// src/services/run-tracking-service.d.ts
export interface ExternalActivity {
    id: string;
    source: "strava" | "garmin" | "apple_health" | "google_fit";
    name: string;
    startTime: number;
    distance: number;
    duration: number;
    polyline?: string;
    averageSpeed: number;
    maxSpeed?: number;
    elevationGain?: number;
    sourceUrl?: string;
}